LKE-9265: Adapt Alert column to support currency type (#622)
* LKE-9265: Adapt Alert column to support currency type

* fix type

* add options to caseColumn

* cleaning

* move currency to common

* rename options field

* add CONSTRAINT_VIOLATION error
